+++ BulletSharp/Collision/GImpact/BoxCollision.cs
@@ -125,7 +125,7 @@ namespace BulletSharp
 			Native = btAABB_new2(ref v1, ref v2, ref v3);
 		}
 
-		public Aabb(Vector3 v1, Vector3 v2, Vector3 v3, float margin)
+		public Aabb(Vector3 v1, Vector3 v2, Vector3 v3, double margin)
 		{
 			Native = btAABB_new3(ref v1, ref v2, ref v3, margin);
 		}
@@ -135,7 +135,7 @@ namespace BulletSharp
 			Native = btAABB_new4(other.Native);
 		}
 
-		public Aabb(Aabb other, float margin)
+		public Aabb(Aabb other, double margin)
 		{
 			Native = btAABB_new5(other.Native, margin);
 		}
@@ -187,7 +187,7 @@ namespace BulletSharp
 				ref trianglePlane);
 		}
 
-		public void CopyWithMargin(Aabb other, float margin)
+		public void CopyWithMargin(Aabb other, double margin)
 		{
 			btAABB_copy_with_margin(Native, other.Native, margin);
 		}
@@ -207,7 +207,7 @@ namespace BulletSharp
 			return btAABB_has_collision(Native, other.Native);
 		}
 
-		public void IncrementMargin(float margin)
+		public void IncrementMargin(double margin)
 		{
 			btAABB_increment_margin(Native, margin);
 		}
@@ -249,12 +249,12 @@ namespace BulletSharp
 			return btAABB_plane_classify(Native, ref plane);
 		}
 
-		public void ProjectionIntervalRef(ref Vector3 direction, out float vmin, out float vmax)
+		public void ProjectionIntervalRef(ref Vector3 direction, out double vmin, out double vmax)
 		{
 			btAABB_projection_interval(Native, ref direction, out vmin, out vmax);
 		}
 
-		public void ProjectionInterval(Vector3 direction, out float vmin, out float vmax)
+		public void ProjectionInterval(Vector3 direction, out double vmin, out double vmax)
 		{
 			btAABB_projection_interval(Native, ref direction, out vmin, out vmax);
 		}
