+++ BulletSharp/Collision/MultiSphereShape.cs
@@ -7,12 +7,12 @@ namespace BulletSharp
 {
 	public class MultiSphereShape : ConvexInternalAabbCachingShape
 	{
-		public MultiSphereShape(Vector3[] positions, float[] radi)
+		public MultiSphereShape(Vector3[] positions, double[] radi)
 			: base(btMultiSphereShape_new(positions, radi, (radi.Length < positions.Length) ? radi.Length : positions.Length))
 		{
 		}
 
-		public MultiSphereShape(Vector3Array positions, float[] radi)
+		public MultiSphereShape(Vector3Array positions, double[] radi)
 			: base(btMultiSphereShape_new2(positions._native, radi, (radi.Length < positions.Count) ? radi.Length : positions.Count))
 		{
 		}
@@ -24,7 +24,7 @@ namespace BulletSharp
 			return value;
 		}
 
-		public float GetSphereRadius(int index)
+		public double GetSphereRadius(int index)
 		{
 			return btMultiSphereShape_getSphereRadius(Native, index);
 		}
