+++ BulletSharp/Collision/TriangleCallback.cs
@@ -25,7 +25,7 @@ namespace BulletSharp
 
 		private void ProcessTriangleUnmanaged(IntPtr triangle, int partId, int triangleIndex)
 		{
-			float[] triangleData = new float[11];
+			double[] triangleData = new double[11];
 			Marshal.Copy(triangle, triangleData, 0, 11);
 			Vector3 p0 = new Vector3(triangleData[0], triangleData[1], triangleData[2]);
 			Vector3 p1 = new Vector3(triangleData[4], triangleData[5], triangleData[6]);
@@ -75,7 +75,7 @@ namespace BulletSharp
 
 		private void InternalProcessTriangleIndexUnmanaged(IntPtr triangle, int partId, int triangleIndex)
 		{
-			float[] triangleData = new float[11];
+			double[] triangleData = new double[11];
 			Marshal.Copy(triangle, triangleData, 0, 11);
 			Vector3 p0 = new Vector3(triangleData[0], triangleData[1], triangleData[2]);
 			Vector3 p1 = new Vector3(triangleData[4], triangleData[5], triangleData[6]);
