+++ BulletSharp/Dynamics/ConstraintSolver.cs
@@ -36,7 +36,7 @@ namespace BulletSharp
 			btConstraintSolver_reset(Native);
 		}
 		/*
-		public float SolveGroup(CollisionObject bodies, int numBodies, PersistentManifold manifold,
+		public double SolveGroup(CollisionObject bodies, int numBodies, PersistentManifold manifold,
 			int numManifolds, TypedConstraint constraints, int numConstraints, ContactSolverInfo info,
 			IDebugDraw debugDrawer, Dispatcher dispatcher)
 		{
