+++ BulletSharp/Dynamics/Featherstone/MultiBodyJointLimitConstraint.cs
@@ -4,8 +4,8 @@ namespace BulletSharp
 {
 	public class MultiBodyJointLimitConstraint : MultiBodyConstraint
 	{
-		public MultiBodyJointLimitConstraint(MultiBody body, int link, float lower,
-			float upper)
+		public MultiBodyJointLimitConstraint(MultiBody body, int link, double lower,
+			double upper)
 			: base(btMultiBodyJointLimitConstraint_new(body.Native, link, lower,
 				upper), body, body)
 		{
