+++ BulletSharp/Dynamics/Featherstone/MultiBodyJointMotor.cs
@@ -4,21 +4,21 @@ namespace BulletSharp
 {
 	public class MultiBodyJointMotor : MultiBodyConstraint
 	{
-		public MultiBodyJointMotor(MultiBody body, int link, float desiredVelocity,
-			float maxMotorImpulse)
+		public MultiBodyJointMotor(MultiBody body, int link, double desiredVelocity,
+			double maxMotorImpulse)
 			: base(btMultiBodyJointMotor_new(body.Native, link, desiredVelocity,
 				maxMotorImpulse), body, body)
 		{
 		}
 
-		public MultiBodyJointMotor(MultiBody body, int link, int linkDoF, float desiredVelocity,
-			float maxMotorImpulse)
+		public MultiBodyJointMotor(MultiBody body, int link, int linkDoF, double desiredVelocity,
+			double maxMotorImpulse)
 			: base(btMultiBodyJointMotor_new2(body.Native, link, linkDoF, desiredVelocity,
 				maxMotorImpulse), body, body)
 		{
 		}
 
-		public void SetVelocityTarget(float velTarget)
+		public void SetVelocityTarget(double velTarget)
 		{
 			btMultiBodyJointMotor_setVelocityTarget(Native, velTarget);
 		}
