+++ BulletSharp/Dynamics/ICharacterController.cs
@@ -9,11 +9,11 @@ namespace BulletSharp
 
         void Jump();
         void Jump(Vector3 dir);
-        void PlayerStep(CollisionWorld collisionWorld, float deltaTime);
+        void PlayerStep(CollisionWorld collisionWorld, double deltaTime);
         void PreStep(CollisionWorld collisionWorld);
         void Reset(CollisionWorld collisionWorld);
         void SetUpInterpolate(bool value);
-        void SetVelocityForTimeInterval(ref Vector3 velocity, float timeInterval);
+        void SetVelocityForTimeInterval(ref Vector3 velocity, double timeInterval);
         void SetWalkDirection(ref Vector3 walkDirection);
         void Warp(ref Vector3 origin);
 	}
